backend/app/requirements.txt
anyio==3.6.2
bcrypt==4.0.1
fastapi==0.95.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
postgis==1.0.4
prisma==0.8.2
pyasn1==0.4.8
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
PyJWT==2.6.0
python-dotenv==1.0.0
starlette==0.26.1
stripe==5.2.0
typing_extensions==4.5.0
urllib3==1.26.15
webargs==8.2.0
Werkzeug==2.2.3
uvicorn==0.21.1
exponent-server-sdk==2.0.0
